ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics>=8.0.196
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
